+++ pathfinder-toolkit/src/com/lateensoft/pathfinder/toolkit/repository/PTStorable.java
@@ -1,6 +1,6 @@
 package com.lateensoft.pathfinder.toolkit.repository;
 
 public interface PTStorable {
-	public int getID();
-	public int getCharacterID();
+	public long getID();
+	public long getCharacterID();
 }
